Fix bugs for PHP 7.2+ and GLPI 9.5.4+
PHP 7.2+ throws a warning if the count() function is called with non-array items, so the reports code was updated to check if $tab is an array before counting it, or returns 0 if it's not an array.

The function num_fields was deprecated with GLPI 9.5.4 and was throwing a warning into GLPI debug mode.

The "Counter records by printer" was updated to include the "Serial", "Inventory/Asset Number" and "Location" of the printer, so now it's matches PrintWayy reports style and makes it easier to identify each printer. The default filter date range was changed from 1 year to 1 month to avoid confusion for people that do tenancy of printers and pay monthly invoices, e.g. it was easier to select 01-02-2020 to 28-02-2021 (notice the Start date is 2020) instead of 01-02-2021 to 28-02-2021 (now a month was selected).